# swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Quick Ask
  description: This is the first professional api project we createing a full-stack army. this is a simple blog api application where anyone can create anew account, login to their system and create article that will be visible publicly.
servers:
  - url: http://localhost:4000/api/v1
    description: Development
...
paths:
...
  "/users":
    get:
      security:
        - bearerAuth: []
      tags:
        - User
      description: Retrive All Users
      parameters:
        - $ref: "#/components/parameters/pageParams"
        - $ref: "#/components/parameters/limitParams"
        - $ref: "#/components/parameters/sortByParams"
        - $ref: "#/components/parameters/sortTypeParams"
        - $ref: "#/components/parameters/searchParams"
      responses:
        200:
          description: Get All User
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Get All Users
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/UserDTO"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                  link:
                    type: object
                    properties:
                      self:
                        type: string
                        example: "/users"
                      nextPage:
                        type: string
                        example: "/users?page=1&limit=10"
                      prevPage:
                        type: string
                        example: "/users?page=3&limit=10"
        401:
          $ref: "#/components/responses/401"
        500:
          $ref: "#/components/responses/500"
    post:
      security:
        - bearerAuth: []
      tags: [User]
      description: Create a user Account
      requestBody:
        content:
          "application/json":
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Aysha Mone
                email:
                  type: string
                  example: mony@gmail.com
                password:
                  type: string
                  format: password
                  example: pass1234
                avatar:
                  type: string
                  example: ""
                role:
                  type: string
                  enum:
                    - user
                    - admin
                  example: user
                status:
                  type: string
                  enum:
                    - public
                    - private
                  example: public
              required: [name, email, password]
      responses:
        201:
          description: Create a new user
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 201
                  message:
                    type: string
                    example: Create a user Successful
                  data:
                    $ref: "#/components/schemas/UserDTO"
        400:
          $ref: "#/components/responses/400_field"
        401:
          $ref: "#/components/responses/401"
        500:
          $ref: "#/components/responses/500"
  "/users/{id}":
    get:
      security:
        - bearerAuth: []
      tags:
        - User
      description: Single a user document
      parameters:
        - $ref: "#/components/parameters/idParamsUser"
      responses:
        200:
          description: Single user document
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Update a user Successfully
                  data:
                    $ref: "#/components/schemas/UserDTO"
        404:
          $ref: "#/components/responses/404"
        400:
          $ref: "#/components/responses/400_field"
        500:
          $ref: "#/components/responses/500"
    patch:
      security:
        - bearerAuth: []
      tags:
        - User
      description: Update a user document
      parameters:
        - $ref: "#/components/parameters/idParamsUser"
      requestBody:
        content:
          "application/json":
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Assaludllah Al Labib
                email:
                  type: string
                  example: labib@gmail.com
                role:
                  type: string
                  enum:
                    - admin
                    - user
                  example: user
                status:
                  type: string
                  enum: [public, private]
                  example: public
                avatar:
                  type: string
                  format: url
                  example: ""
      responses:
        200:
          description: Update a user document Successfully
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Update a user Successfully
                  data:
                    $ref: "#/components/schemas/UserDTO"
        404:
          $ref: "#/components/responses/404"
        400:
          $ref: "#/components/responses/400_field"
        500:
          $ref: "#/components/responses/500"
    delete:
      security:
        - bearerAuth: []
      tags: [User]
      description: Delete a user
      parameters:
        - $ref: "#/components/parameters/idParamsUser"
      responses:
        204:
          description: User Delete Successfully
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 204
                  message:
                    type: string
                    example: User Deleted Successfully
        404:
          $ref: "#/components/responses/404"
        500:
          $ref: "#/components/responses/500"
  "/users/{id}/password":
    patch:
      security:
        - bearerAuth: []
      tags: [User]
      description: Change User Password
      parameters:
        - $ref: "#/components/parameters/idParamsUser"
      requestBody:
        content:
          "application/json":
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: 123abce
      responses:
        200:
          description: Change User password
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Password change Successfully
        404:
          $ref: "#/components/responses/404"
        401:
          $ref: "#/components/responses/401"
        500:
          $ref: "#/components/responses/500"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    "500":
      description: Internal Server error
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 500
              error:
                type: string
                example: Internal Server Error
              message:
                type: string
                example: We are sorry for the inconvenience. Please Try again later
...
    "400_field":
      description: Bad Request
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 400
              message:
                type: string
                example: Bad Request
              error:
                type: array
                items:
                  type: object
                  properties:
                    message:
                      type: string
                      example: name is require
                    filed:
                      type: string
                      example: name
    "401":
      description: Unauthorize
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 401
              message:
                type: string
                example: Your are not permitted this operation
    "404":
      description: Resources Not Found
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 404
              message:
                type: string
                example: Resources not Found
  parameters:
    pageParams:
      in: query
      name: page
      description: Current page number
      schema:
        type: integer
        format: int32
        example: 1
    limitParams:
      in: query
      name: limit
      description: the number of object should be return
      schema:
        type: integer
        format: int32
        example: 10
...
    idParamsUser:
      in: path
      name: id
      required: true
      description: User Id
      schema:
        type: string
        example: "1234567abc"
    sortTypeParams:
      in: query
      name: sort_type
      description: Type of sort (asc or dsc)
      schema:
        type: string
        enum:
          - asc
          - dsc
        example: dsc
    sortByParams:
      in: query
      name: sort_by
      description: Sort Key (updateAt or title)
      schema:
        type: string
        example: updatedAt
    searchParams:
      in: query
      name: search
      description: Search By User Name
      schema:
        type: string
        example: User Name
  schemas:
    Timestamp:
      type: object
      properties:
        createdAt:
          type: string
          example: "2023-08-25T06:45:12.159Z"
        updatedAt:
          type: string
          example: "2023-08-25T06:46:08.634Z"
...
    UserDTO:
      allOf:
        - type: object
          properties:
            id:
              type: string
              example: user12344
            name:
              type: string
              example: Abdullah Al Zarif
            email:
              type: string
              example: zarif@gmail.com
            avatar:
              type: string
              format: url
              example: ""
            role:
              type: string
              enum: [admin, user]
              example: user
            status:
              type: string
              enum:
                - public
                  private
              example: public
            link:
              type: string
              example: users/user123
        - $ref: "#/components/schemas/Timestamp"
        - required:
            - name
            - email
            - id
    Pagination:
      type: object
      properties:
        page:
          type: integer
          format: int32
          example: 2
        limit:
          type: integer
          format: int32
          example: 10
        next:
          type: integer
          format: int32
          example: 3
        prev:
          type: integer
          format: int32
          example: 1
        totalItems:
          type: integer
          format: int32
          example: 25
        totalPage:
          type: integer
          format: int32
          example: 3
      required: [page, limit, totalItems, totalPage]